fix: clear core compilation warnings and enhance string safety
- Resolved -Wswitch warnings in dashboard by handling all new event types
- Replaced dangerous strncpy with snprintf in config and server logic
- Fixed -Wsign-compare and -Wformat issues in source and test files
- Cleaned up multiple unused variable warnings in unit tests
- Verified functional integrity with full-test suite

+++ src/tsa_source.c
@@ -99,13 +99,13 @@ static void pcap_packet_callback(u_char* user, const struct pcap_pkthdr* h, cons
         return;
     }
 
-    if (h->caplen < offset + 20 + 8) return;
+    if (h->caplen < (uint32_t)(offset + 20 + 8)) return;
 
     struct iphdr* ip = (struct iphdr*)(pkt + offset);
     if (ip->version != 4 || ip->protocol != IPPROTO_UDP) return;
 
     int ip_hdr_len = ip->ihl * 4;
-    if (h->caplen < offset + ip_hdr_len + 8) return;
+    if (h->caplen < (uint32_t)(offset + ip_hdr_len + 8)) return;
 
     struct udphdr* udp = (struct udphdr*)(pkt + offset + ip_hdr_len);
     uint8_t* payload = (uint8_t*)udp + 8;
